なんか今度は地図側の枠が太い気がする。計算は合ってると思うんだけどな。

+++ selfdrive/ui/qt/onroad.cc
@@ -148,7 +148,7 @@ void OnroadWindow::paintEvent(QPaintEvent *event) {
   if (isMapVisible() && navDisabled) {
     int w;
     if(this->mb_width_rate != 0){
-      w = (width() - bdr_s * 2) * this->mb_width_rate + bdr_s;
+      w = (int)((width() - bdr_s * 2) * this->mb_width_rate) + bdr_s;
     } else {
       w = (width() - bdr_s * 2) /2 + bdr_s;
     }
